Ensure photos and observation_photos get returned even if there are flags.
For some reason, without this, getting a single observation with no photos but
one flag would return a response with no photos attribute, which was causing
problems b/c our obs detail page seems to expect an empty array if there are no
photos. I don't really understand why that was happening, but this seems to fix
it.