Set CORS headers when req val fails; populate query with body on override
Access Control headers were not getting set when Open API request validation was
failing, which meant clients like Chrome were not getting useful error messages.

HTTP override was not quite working for endpoints like taxa/autocomplete that
were expecting content in req.query, so this populates that with the content in
req.body.