Including obs in trusted collection projects was not working for umbrellas
Specifically, if you trusted an umbrella but none of the projects in that
umbrella, your hidden coordinates were not visible to the umbrella admins and
your obs did not appear to be in the umbrella. While this will probably cause
some confusion b/c obs will appear to be in the umbrella without being in any of
the constituent projects, it will allow trust to be given once to an umbrella
instead of repeatedly for constituent projects, making data management
significantly easier for project managers running large umbrellas.