Don't stringify null when converting a POST body to a GET query
This conversion seems like something we need to do get coersion to work, but we
were converting null to {}, which was causing a validation error when the param
was supposed to be an integer or something. I thought this was because it was
objecting to a null value, but it's actually ok with a null, just not a blank
object.