Ensure description and other fields get returned for /v2/users/:id
The description attribute is not in the ES index, so our fixtures were
actually incorrect. This fixes the fixtures and adds a new implementation
of /v2/users/:id that loads data from the db and from ES every time, and can
return all the in the documented response schema... which itself was wildly
inaccurate as it included all of the private info returned by /v2/users/me,
so I trimmed that down and made a separate schema for that.