Add /v2/log endpoint so clients can write log entries
This is mainly to support monitoring and crash logging in our own client
applications. It builds on the abilities of /v1/log but tries to support
non-error logging as well. Some of the attributes it writes should probably
be added to the index, but for now I've added them to the `qry` object.